+++ tests/IntegrationTest/DefaultEntriesTest.php
@@ -70,8 +70,9 @@ class DefaultEntriesTest extends BaseContainerTest
         $builder->wrapContainer($wrapperContainer);
         $container = $builder->build();
 
-        $this->assertSame($wrapperContainer, $container->get(Container::class));
         $this->assertSame($wrapperContainer, $container->get(ContainerInterface::class));
+        // These entries must point to the PHP-DI instance because the wrapper container is type-hinted as `ContainerInterface` only
+        $this->assertSame($container, $container->get(Container::class));
         $this->assertSame($container, $container->get(InvokerInterface::class));
         $this->assertSame($container, $container->get(FactoryInterface::class));
     }